extract socket getter
