+++ cmd/dbx-dev/actions.go
@@ -1,13 +1,11 @@
 package dbxdev
 
 import (
-	"context"
 	"crypto/sha256"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -22,17 +20,7 @@ import (
 // and then triggers a refresh of pup list.
 func pupActionCmd(id, action string) tea.Cmd {
 	return func() tea.Msg {
-		socketPath := os.Getenv("DBX_SOCKET")
-		if socketPath == "" {
-			socketPath = filepath.Join(os.Getenv("HOME"), "data", "dbx-socket")
-		}
-
-		tr := &http.Transport{
-			DialContext: func(ctx context.Context, _, _ string) (net.Conn, error) {
-				return net.Dial("unix", socketPath)
-			},
-		}
-		client := &http.Client{Transport: tr, Timeout: 5 * time.Second}
+		client := getSocketClient()
 
 		url := "http://dogeboxd/pup/" + id + "/" + action
 		req, err := http.NewRequest(http.MethodPost, url, nil)
